@@ -1,4 +1,4 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 5.10
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -13,38 +13,31 @@ let package = Package(
       name: "mew-wallet-ios-walletconnect",
       targets: ["mew-wallet-ios-walletconnect"]),
     .library(
-      name: "mew-wallet-ios-walletconnect-v1",
-      targets: ["mew-wallet-ios-walletconnect-v1"]),
-    .library(
       name: "mew-wallet-ios-walletconnect-v2",
       targets: ["mew-wallet-ios-walletconnect-v2"])
   ],
   dependencies: [
     .package(url: "https://github.com/WalletConnect/WalletConnectSwiftV2", exact: "1.11.0"),
-    .package(url: "https://github.com/krzyzanowskim/CryptoSwift", from: "1.4.0"),
-    .package(url: "https://github.com/daltoniam/Starscream", from: "4.0.0"),
     .package(url: "https://github.com/mewwallet/mew-wallet-ios-logger.git", .upToNextMajor(from: "2.0.0"))
   ],
   targets: [
     .target(
       name: "mew-wallet-ios-walletconnect",
       dependencies: [
-        "mew-wallet-ios-walletconnect-v1",
         "mew-wallet-ios-walletconnect-v2"
       ],
-      path: "Sources/mew-wallet-ios-walletconnect"
-    ),
-    .target(
-      name: "mew-wallet-ios-walletconnect-v1",
-      dependencies: [
-        .product(name: "CryptoSwift", package: "CryptoSwift"),
-        .product(name: "Starscream", package: "Starscream"),
-        .product(name: "mew-wallet-ios-logger", package: "mew-wallet-ios-logger")
+      path: "Sources/mew-wallet-ios-walletconnect",
+      resources: [
+        .copy("Privacy/PrivacyInfo.xcprivacy")
       ],
-      path: "Sources/v1"),
+      swiftSettings: [
+        .enableExperimentalFeature("StrictConcurrency=complete")
+      ]
+    ),
     .target(
       name: "mew-wallet-ios-walletconnect-v2",
       dependencies: [
+        .product(name: "mew-wallet-ios-logger", package: "mew-wallet-ios-logger"),
         .product(name: "WalletConnect", package: "WalletConnectSwiftV2"),
         .product(name: "WalletConnectNetworking", package: "WalletConnectSwiftV2"),
         .product(name: "WalletConnectPush", package: "WalletConnectSwiftV2"),
@@ -53,14 +46,18 @@ let package = Package(
         .product(name: "WalletConnectNotify", package: "WalletConnectSwiftV2"),
         .product(name: "WalletConnectSync", package: "WalletConnectSwiftV2"),
         .product(name: "WalletConnectPairing", package: "WalletConnectSwiftV2"),
-        .product(name: "Starscream", package: "Starscream")
       ],
-      path: "Sources/v2"),
+      path: "Sources/v2",
+      resources: [
+        .copy("Privacy/PrivacyInfo.xcprivacy")
+      ],
+      swiftSettings: [
+        .enableExperimentalFeature("StrictConcurrency=complete")
+      ]),
     .testTarget(
       name: "mew-wallet-ios-walletconnectTests",
       dependencies: [
         "mew-wallet-ios-walletconnect",
-        "mew-wallet-ios-walletconnect-v1"
       ]),
   ]
 )
