@@ -1,10 +1,9 @@
 import Foundation
-import Combine
-import WalletConnectSign
-import WalletConnectNetworking
-import Auth
+@preconcurrency import Combine
+@preconcurrency import WalletConnectSign
+@preconcurrency import Auth
 
-public final class WalletConnectSessionPublisher {
+public final class WalletConnectSessionPublisher: Sendable {
   internal let pushOnSignSubject = PassthroughSubject<PushOnSign, Never>()
   
   /// Your Wallet should allow users to scan a QR code generated by dapps. You are responsible for implementing it on your own. For testing, you can use our test dapp at: https://react-app.walletconnect.com/, which is v2 protocol compliant. Once you derive a URI from the QR code call pair method: try await WallectConnectServiceImpl.instance.pair(uri: uri)
