@@ -6,18 +6,14 @@
 //
 
 import Foundation
-import WalletConnectSign
-import mew_wallet_ios_walletconnect_v1
+@preconcurrency import WalletConnectSign
 import mew_wallet_ios_walletconnect_v2
 
 public enum SessionProposal {
-  case v1(request: WC1.Request, session: WC1.Session)
   case v2(proposal: WC2.Session.Proposal, context: WC2.VerifyContext?)
   
   public var chainIds: [UInt64]? {
     switch self {
-    case .v1(_, let session):
-      return [session.chainId]
     case .v2(let proposal, _):
       guard !proposal.requiredNamespaces.contains(where: { !$0.key.hasPrefix("eip155") }) else { return nil }
       let chainIdsReferences = proposal.requiredNamespaces.flatMap { (key, namespace) -> [String] in
@@ -40,8 +36,6 @@ public enum SessionProposal {
   
   public var optionalChainIds: [UInt64]? {
     switch self {
-    case .v1:
-      return nil
     case .v2(let proposal, _):
       guard let namespace = proposal.optionalNamespaces else { return nil }
       guard !namespace.contains(where: { !$0.key.hasPrefix("eip155") }) else { return nil }
@@ -64,7 +58,7 @@ public enum SessionProposal {
   }
   
   public var redirect: String? {
-    guard case .v2(let proposal, let context) = self else {
+    guard case .v2(let proposal, _ /*let context*/) = self else {
       return nil
     }
     return proposal.proposer.redirect?.native ?? proposal.proposer.redirect?.universal
@@ -74,3 +68,7 @@ public enum SessionProposal {
 // MARK: - SessionProposal + Equatable
 
 extension SessionProposal: Equatable {}
+
+// MARK: - SessionProposal + Sendable
+
+extension SessionProposal: Sendable {}
