@@ -6,16 +6,7 @@
 //
 
 import Foundation
-import mew_wallet_ios_walletconnect_v1
 import mew_wallet_ios_walletconnect_v2
-
-public struct WC1 {
-  public typealias WalletConnectProvider = mew_wallet_ios_walletconnect_v1.WalletConnectProvider
-  public typealias SessionStorage = mew_wallet_ios_walletconnect_v1.SessionStorage
-  public typealias Session = mew_wallet_ios_walletconnect_v1.Session
-  public typealias Request = mew_wallet_ios_walletconnect_v1.JSONRPC.Request
-  public typealias Reason = mew_wallet_ios_walletconnect_v1.Reason
-}
 
 public struct WC2 {
   public typealias WalletConnectProvider = mew_wallet_ios_walletconnect_v2.WalletConnectProvider
@@ -29,4 +20,6 @@ public struct WC2 {
   public typealias SIWECacaoFormatter = mew_wallet_ios_walletconnect_v2.SIWECacaoFormatter
   public typealias APNSEnvironment = mew_wallet_ios_walletconnect_v2.APNSEnvironment
   public typealias PushOnSignRequest = PushOnSign
+  public typealias WebSocketFactory = mew_wallet_ios_walletconnect_v2.WebSocketFactory
+  public typealias WebSocketConnecting = mew_wallet_ios_walletconnect_v2.WebSocketConnecting
 }
