@@ -7,7 +7,6 @@
 
 import os
 import Foundation
-import mew_wallet_ios_walletconnect_v1
 import mew_wallet_ios_walletconnect_v2
 
 public final class WalletConnectProvider {
@@ -18,59 +17,38 @@ public final class WalletConnectProvider {
   /// Query sessions
   /// - Returns: All sessions
   public var sessions: [Session] {
-    let v1: [Session] = WC1.WalletConnectProvider.instance.sessions
-      .map({ .v1(session: $0) })
-    
-    let v2: [Session] = WC2.WalletConnectProvider.instance.sessions
+    return WC2.WalletConnectProvider.instance.sessions
       .map({ .v2(session: $0) })
-    
-    return (v1 + v2).sorted()
+      .sorted()
   }
   
   // Pairing
   public func pair(url: String?) async throws {
     guard let url else { return }
     do {
-      // Try v2 first
+      // Try v2
       try await WC2.WalletConnectProvider.instance.pair(url: url)
     } catch WalletConnectServiceError.invalidPairingURL {
-      do {
-        // If fails - try v1
-        try await WC1.WalletConnectProvider.instance.pair(url: url)
-      } catch {
-        throw error
-      }
+      throw WalletConnectServiceError.invalidPairingURL
     } catch {
       Logger.debug(.rootProvider, "Error: \(error)")
     }
   }
   
-  public func cancel(url: String?) async throws {
-    guard let url else { return }
-    do {
-      // TODO: v2
-      throw WalletConnectServiceError.invalidPairingURL
-    } catch WalletConnectServiceError.invalidPairingURL {
-      do {
-        try await WC1.WalletConnectProvider.instance.cancelPair(url: url)
-      } catch {
-        throw error
-      }
-    }
-  }
-  
-  public func configure(projectId: String, groupIdentifier: String, notifications: (pushHost: String?, environment: WC2.APNSEnvironment)? = nil, metadata: WC2.AppMetadata, storage: WC1.SessionStorage, provider: WC2.CryptoProvider) {
+  public func configure(
+    projectId: String,
+    groupIdentifier: String,
+    notifications: (pushHost: String?, environment: WC2.APNSEnvironment)? = nil,
+    metadata: WC2.AppMetadata,
+    provider: WC2.CryptoProvider,
+    socketFactory: any WC2.WebSocketFactory
+  ) {
     // Configure v2
-    WC2.WalletConnectProvider.instance.configure(projectId: projectId, groupIdentifier: groupIdentifier, notifications: notifications, metadata: metadata, cryptoProvider: provider)
-    // Configure v1
-    let metadata = WC1.Session.AppMetadata(name: metadata.name, url: metadata.url, description: metadata.description, icons: metadata.icons)
-    WC1.WalletConnectProvider.instance.configure(metadata: metadata, storage: storage)
+    WC2.WalletConnectProvider.instance.configure(projectId: projectId, groupIdentifier: groupIdentifier, notifications: notifications, metadata: metadata, cryptoProvider: provider, socketFactory: socketFactory)
   }
   
-  public func approve<T: Codable>(request: Request, result: T) async throws {
+  public func approve<T: Codable & Sendable>(request: Request, result: T) async throws {
     switch request {
-    case .v1(let request, let session):
-      try await WC1.WalletConnectProvider.instance.approve(request: request, for: session, result: result)
     case .v2(let request, _, _):
       try await WC2.WalletConnectProvider.instance.approve(request: request, result: result)
     }
@@ -78,8 +56,6 @@ public final class WalletConnectProvider {
   
   public func reject(request: Request) async throws {
     switch request {
-    case .v1(let request, let session):
-      try await WC1.WalletConnectProvider.instance.reject(request: request, for: session)
     case .v2(let request, _, _):
       try await WC2.WalletConnectProvider.instance.reject(request: request)
     }
@@ -93,8 +69,6 @@ public final class WalletConnectProvider {
   ///   - topic: Session that you want to delete
   public func disconnect(session: Session) async throws {
     switch session {
-    case .v1(let session):
-      try await WC1.WalletConnectProvider.instance.disconnect(session: session)
     case .v2(let session):
       try await WC2.WalletConnectProvider.instance.disconnect(session: session)
     }
@@ -106,8 +80,6 @@ public final class WalletConnectProvider {
   ///   - reason: Reason why the session proposal has been rejected. Conforms to CAIP25.
   public func reject(proposal: SessionProposal, reason: RejectionReason) async throws {
     switch proposal {
-    case .v1(let request, let session):
-      try await WC1.WalletConnectProvider.instance.reject(proposal: request, for: session)
     case .v2(let proposal, _):
       try await WC2.WalletConnectProvider.instance.reject(proposalId: proposal.id, reason: reason)
     }
@@ -115,9 +87,6 @@ public final class WalletConnectProvider {
   
   public func approve(proposal: SessionProposal, accounts: [String], chains: [UInt64], supportedMethods: Set<String> = []) async throws {
     switch proposal {
-    case .v1(let request, let session):
-      guard let chain = chains.first else { throw WalletConnectServiceError.badParameters }
-      try await WC1.WalletConnectProvider.instance.approve(proposal: request, for: session, result: accounts, chainId: chain)
     case .v2(let proposal, _):
       try await WC2.WalletConnectProvider.instance.approve(proposal: proposal, chains: chains, accounts: accounts, supportedMethods: supportedMethods)
       
@@ -126,8 +95,6 @@ public final class WalletConnectProvider {
   
   public func update(session: Session, chainId: UInt64?, accounts: [String]) async throws {
     switch session {
-    case .v1(let session):
-      try await WC1.WalletConnectProvider.instance.update(session: session, chainId: chainId, accounts: accounts)
     case .v2(let session):
       try await WC2.WalletConnectProvider.instance.update(session: session, chainId: chainId, accounts: accounts)
       break
@@ -169,7 +136,6 @@ public final class WalletConnectProvider {
   }
   
   public func reset() async {
-    await WC1.WalletConnectProvider.instance.reset()
     await WC2.WalletConnectProvider.instance.reset()
   }
   
