@@ -10,6 +10,6 @@ import Foundation
 import mew_wallet_ios_logger
 
 extension Logger.System {
-  public static var networking     = Logger.System.with(subsystem: "com.myetherwallet.mewwallet.walletconnect.v2", category: "v2: networking")
-  public static var provider       = Logger.System.with(subsystem: "com.myetherwallet.mewwallet.walletconnect.v2", category: "v2: provider")
+  public static let networking     = Logger.System.with(subsystem: "com.myetherwallet.mewwallet.walletconnect.v2", category: "v2: networking")
+  public static let provider       = Logger.System.with(subsystem: "com.myetherwallet.mewwallet.walletconnect.v2", category: "v2: provider")
 }
