@@ -1,9 +1,8 @@
 import XCTest
 @testable import mew_wallet_ios_walletconnect
-@testable import mew_wallet_ios_walletconnect_v1
 
 final class mew_wallet_ios_walletconnectTests: XCTestCase {
-  func testWCURI_v1() throws {
-    
+  func testSomething() throws {
+    // :)
   }
 }
